+++ src/lib/DataTableFilterRow/DataTableFilterRow.tsx
@@ -3,28 +3,31 @@ import { Input } from "reactstrap";
 import { DataTableColumnDescription, DataTableActions, Filters, FilterTranslations } from "../DataTable/DataTableInterfaces";
 import { ColumnFilterType, ActionsPosition } from "../DataTable/DataTableTypes";
 import { ActionsHeaderFilterCell } from "../DataTable/Actions/ActionsHeaderFilterCell";
+import { getDeepValue } from "../Utils/DeepValue";
 
-interface DataTableFilterRowProps<T> {
+interface DataTableFilterRowProps<T, TFilter> {
   columns: DataTableColumnDescription<T>[];
   actions?: DataTableActions<T>;
   actionsPosition?: ActionsPosition;
   translations: FilterTranslations;
   filterPossible: boolean;
+  predefinedFilter?: TFilter;
   getFilterRefs(): Filters;
   setFilterRef(filterName: string, ref: HTMLInputElement): void;
   onSearch(): void;
 }
 
-export function DataTableFilterRow<T>({
+export function DataTableFilterRow<T, TFilter>({
   columns,
   actions,
   actionsPosition,
   filterPossible = true,
   translations,
+  predefinedFilter,
   getFilterRefs,
   setFilterRef,
   onSearch,
-}: DataTableFilterRowProps<T>) {
+}: DataTableFilterRowProps<T, TFilter>) {
   if (!filterPossible || columns.filter((column) => column.filter).length <= 0) return <React.Fragment />;
   return (
     <tr>
@@ -38,6 +41,7 @@ export function DataTableFilterRow<T>({
               bsSize="sm"
               id={`filter-${column.dataField}`}
               innerRef={(ref) => ref && ref instanceof HTMLInputElement && setFilterRef(column.dataField, ref)}
+              defaultValue={predefinedFilter ? getDeepValue(predefinedFilter, column.dataField) : undefined}
               onKeyDown={(e) => {
                 if (e.key === "Enter") onSearch();
               }}
@@ -48,6 +52,7 @@ export function DataTableFilterRow<T>({
               bsSize="sm"
               innerRef={(ref) => ref && ref instanceof HTMLSelectElement && setFilterRef(column.dataField, ref as HTMLInputElement)}
               type="select"
+              defaultValue={predefinedFilter ? getDeepValue(predefinedFilter, column.dataField) : undefined}
               onKeyDown={(e) => {
                 if (e.key === "Enter") onSearch();
               }}